Fix issue with android versions below 23
